requirements.txt
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
stripe==8.4.0
python-dotenv==1.0.1
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
prophet==1.1.5
xgboost==2.0.3
tensorflow==2.15.0
openai==1.12.0
langchain==0.1.7
pydantic==2.6.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
reportlab==4.0.9
openpyxl==3.1.2
xlrd==2.0.1
python-magic==0.4.27
cryptography==42.0.2
PyJWT==2.8.0
oauthlib==3.2.2
python3-openid==3.2.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
django-extensions==3.2.3
Werkzeug==3.0.1
